migration INT to BIGINT type changed for delta

+++ internal/server/database/migrations/000001_init.up.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS public.metrics
 (
     id VARCHAR PRIMARY KEY,
     type VARCHAR not null,
-    delta INT null,
+    delta BIGINT null,
     value DOUBLE PRECISION null
 );
